requirements.txt
pysam
pyvcf
matplotlib
pandas
scipy
seaborn
progressbar
